Support custom formatters for Bugsnag
It also refactor the way default config a provided by the notifier
classes. I allow to get a option's Hash filled with the default during
the configuration step:

config.notifier :bugsnag do |options|
  # Here options[:formatter_class] is set to the default
end

It also refactor the way we call Bugsnag in order to use a compatible
syntax with Bugsnag 6 reports (rather than notifications). The tests are
updated in that direction in order to avoid Bugsnag 6 to pass the tests
as it does today in #11.

Since rubocop was updated, some new offenses has been fixed.